+++ nodes/current-state/current-state.js
@@ -48,7 +48,7 @@ module.exports = function (RED) {
         }
 
         /* eslint-disable camelcase */
-        async onInput({ parsedMessage, message }) {
+        onInput({ parsedMessage, message }) {
             const config = this.nodeConfig;
             const entityId = RenderTemplate(
                 config.blockInputOverrides === true
@@ -64,9 +64,7 @@ module.exports = function (RED) {
                 return;
             }
 
-            const entity = await config.server.homeAssistant.getStates(
-                entityId
-            );
+            const entity = config.server.homeAssistant.getStates(entityId);
             if (!entity) {
                 this.node.error(
                     `Entity could not be found in cache for entity_id: ${entityId}`,
@@ -79,13 +77,18 @@ module.exports = function (RED) {
                 Date.now() - new Date(entity.last_changed).getTime();
 
             // Convert and save original state if needed
-            if (config.state_type !== 'str') {
-                entity.original_state = entity.state;
-                entity.state = this.getCastValue(
-                    config.state_type,
-                    entity.state
-                );
-            }
+            this.castState(entity, config.state_type);
+
+            const isIfState = this.getComparatorResult(
+                config.halt_if_compare,
+                config.halt_if,
+                entity.state,
+                config.halt_if_type,
+                {
+                    message,
+                    entity,
+                }
+            );
 
             // default switch to true if undefined (backward compatibility)
             message.topic =
@@ -106,20 +109,6 @@ module.exports = function (RED) {
                 config.entity_location,
                 message
             );
-
-            const isIfState = await this.getComparatorResult(
-                config.halt_if_compare,
-                config.halt_if,
-                entity.state,
-                config.halt_if_type,
-                {
-                    message,
-                    entity,
-                }
-            ).catch((e) => {
-                this.setStatusFailed('Error');
-                this.node.error(e.message, message);
-            });
 
             // Handle version 0 'halt if' outputs
             if (config.version < 1) {